admin/requirements.txt
django>=6.0
google-cloud-firestore>=2.22.0
google-cloud-aiplatform>=1.86.0
google-cloud-run>=0.10.0
google-genai>=0.2.0
gunicorn>=25.1.0
python-dotenv>=1.2.2
dj-database-url>=2.3.0
psycopg2-binary>=2.9.10